Fix bug in acknowledget_interrupt
Before we were passing the address of the ct2_in structure.
Python does not support ioctl calls with > 32bits in the arg parameter.
This lead to random errors depending on the address being passed.

Now we pass the ct2_in and ask ioctl to interpret it as a writable buffer.